+++ assets/js/dashboard/components/Profile/Security.vue
@@ -5,15 +5,23 @@
             prepend-icon="lock"
             type="password"
             :rules="passwordRules"
-            label="Password"
+            label="Current Password"
             required
         ></v-text-field>
         <v-text-field
-            v-model="passwordConfirm"
+            v-model="newPassword"
             prepend-icon="lock"
             type="password"
-            :rules="passwordConfirmRules"
-            label="Password Confirmation"
+            :rules="newPasswordRules"
+            label="New Password"
+            required
+        ></v-text-field>
+        <v-text-field
+            v-model="newPasswordConfirm"
+            prepend-icon="lock"
+            type="password"
+            :rules="newPasswordConfirmRules"
+            label="New Password Confirmation"
             required
         ></v-text-field>
         <v-btn color="success" @click="save">
@@ -28,16 +36,18 @@ import axios from "axios";
 export default {
     data: () => ({
         password: "",
-        passwordRules: [v => !!v || "Please provide your password"],
-        passwordConfirm: ""
+        passwordRules: [v => !!v || "Please provide your current password"],
+        newPassword: "",
+        newPasswordRules: [v => !!v || "Please provide your new password"],
+        newPasswordConfirm: ""
     }),
     computed: {
-        passwordConfirmRules() {
+        newPasswordConfirmRules() {
             return [
+                v => !!v || "Please confirm your password",
                 () =>
-                    this.password === this.passwordConfirm ||
-                    "Passwords does not match",
-                v => !!v || "Please confirm your password"
+                    this.newPassword === this.newPasswordConfirm ||
+                    "Passwords does not match"
             ];
         }
     },
@@ -45,8 +55,15 @@ export default {
         console.log("mounted");
     },
     methods: {
-        save() {
+        async save() {
             if (this.$refs.form.validate()) {
+                let data = {
+                        password: this.password,
+                        newPassword: this.newPassword
+                    },
+                    response = await axios.post("/api/change-password", data);
+
+                this.$toast.success("Password changed");
             }
         }
     }